Refactor authentication logic by moving routes to a separate file Renamed index.js to server.js (and simplified the file)
